requirements.txt
django==5.0.6
django-crispy-forms==2.1
django-extensions==3.2.3
crispy-bootstrap3==2024.1
pyyaml==6.0.2
